requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/pre-release
--extra-index-url https://storage.openvinotoolkit.org/simple/wheels/nightly
openvino_genai~=2024.6.0.0.dev
openvino-tokenizers~=2024.6.0.0.dev
# optimum is pinned because of Whisper is changed since transformers 4.47 (https://github.com/huggingface/optimum-intel/pull/1088)
optimum-intel @ git+https://github.com/huggingface/optimum-intel.git@753f84db6e0966580eb9eaa74a808213be730631
numpy<2.0.0 #; sys_platform == 'darwin'
einops==0.8.0  # For Qwen
transformers_stream_generator==0.0.5  # For Qwen
diffusers==0.31.0 # For image generation pipelines
timm==1.0.11  # For exporting InternVL2
torchvision  # For visual language models
transformers>=4.43 # For Whisper
gguf
torch 
huggingface_hub